+++ src/ast.rs
@@ -131,7 +131,6 @@ pub enum Word {
         in_macro_arg: bool,
     },
     Strand(Vec<Sp<Word>>),
-    Undertied(Vec<Sp<Word>>),
     Array(Arr),
     Func(Func),
     Pack(FunctionPack),
@@ -163,9 +162,6 @@ impl PartialEq for Word {
             (Self::MultilineFormatString(a), Self::MultilineFormatString(b)) => a == b,
             (Self::Ref(a), Self::Ref(b)) => a == b,
             (Self::Strand(a), Self::Strand(b)) => {
-                a.iter().map(|w| &w.value).eq(b.iter().map(|w| &w.value))
-            }
-            (Self::Undertied(a), Self::Undertied(b)) => {
                 a.iter().map(|w| &w.value).eq(b.iter().map(|w| &w.value))
             }
             (Self::Array(a), Self::Array(b)) => a.lines.iter().flatten().map(|w| &w.value).eq(b
@@ -263,7 +259,6 @@ impl fmt::Debug for Word {
             }
             Word::Array(arr) => arr.fmt(f),
             Word::Strand(items) => write!(f, "strand({items:?})"),
-            Word::Undertied(items) => write!(f, "undertie({items:?})"),
             Word::Func(func) => func.fmt(f),
             Word::Pack(pack) => pack.fmt(f),
             Word::Primitive(prim) => prim.fmt(f),